Consolidate reasonable cards into one place.
Putting in cardutil since carddb uses it. Using the version from carddb
since it has more conditions.
